Started on the CanCan adapater

+++ lib/active_admin/views/index_as_table.rb
@@ -152,13 +152,13 @@ module ActiveAdmin
           }.merge(options)
           column options[:name] do |resource|
             links = ''.html_safe
-            if controller.action_methods.include?('show')
+            if controller.action_methods.include?('show') && authorized?(ActiveAdmin::Auth::READ, resource)
               links << link_to(I18n.t('active_admin.view'), resource_path(resource), :class => "member_link view_link")
             end
-            if controller.action_methods.include?('edit')
+            if controller.action_methods.include?('edit') && authorized?(ActiveAdmin::Auth::UPDATE, resource)
               links << link_to(I18n.t('active_admin.edit'), edit_resource_path(resource), :class => "member_link edit_link")
             end
-            if controller.action_methods.include?('destroy')
+            if controller.action_methods.include?('destroy') && authorized?(ActiveAdmin::Auth::DESTROY, resource)
               links << link_to(I18n.t('active_admin.delete'), resource_path(resource), :method => :delete, :data => {:confirm => I18n.t('active_admin.delete_confirmation')}, :class => "member_link delete_link")
             end
             links
